refactor : goalMoney 타입 number로 통일
